requirements.txt
whisperx
whisper
pyannote.audio
langchain-core
langgraph
langfuse
python-dotenv
pydantic>=2.0
fastapi
uvicorn
gradio